Rework import/export structure
