+++ packages/design-system/src/components/Card/index.ts
@@ -14,7 +14,5 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 // =============================================================================
-declare module "*.svg" {
-  const content: string;
-  export default content;
-}
+
+export * from "./Card.styles";
