+++ packages/design-system/src/components/ChartWrapper/ChartWrapper.stories.mdx
@@ -21,7 +21,7 @@ import { Canvas, Meta, Story, ArgsTable } from "@storybook/addon-docs/blocks";
 import { cityTemperature, exoplanets, letterFrequency } from "@visx/mock-data";
 import { OrdinalFrame, XYFrame } from "semiotic";
 import { palette } from "../../styles";
-import ChartWrapper from "./ChartWrapper";
+import { ChartWrapper } from "./ChartWrapper";
 
 <Meta
   title="Design System/Charts"
