+++ packages/design-system/src/components/Tabs/index.ts
