+++ packages/design-system/src/components/Typography/index.ts
@@ -14,6 +14,5 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 // =============================================================================
-
-export { CardSection, Card } from "./Card.styles";
-export type { CardProps } from "./Card.styles";
+export * from "./Headings";
+export * from "./Link";
