+++ packages/design-system/src/index.tsx
@@ -14,63 +14,21 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 // =============================================================================
-import { GlobalStyle } from "./styles/GlobalStyle";
-import { Button, ButtonKind, ButtonProps } from "./components/Button/Button";
-import Dropdown from "./components/Dropdown";
-import { Card, CardSection } from "./components/Card/Card";
-import { ErrorPage } from "./components/ErrorPage/ErrorPage";
-import { Header } from "./components/Header/Header";
-import { Icon } from "./components/Icon/Icon";
-import { IconSVG } from "./components/Icon/IconSVG";
-import { Need } from "./components/Need/Need";
-import { Search } from "./components/Search/Search";
-import { fonts, palette, spacing, zindex } from "./styles";
-
-import {
-  TitleXXL,
-  TitleXL,
-  H1,
-  H2,
-  H3,
-  H4,
-} from "./components/Typography/Headings";
-
 import Assets from "./assets";
-import { Modal, ModalHeading, ModalProps } from "./components/Modal/Modal";
-import { Link } from "./components/Typography/Link";
-import ChartWrapper from "./components/ChartWrapper/ChartWrapper";
-import { NeedState } from "./components/Need/Need.types";
 
-export {
-  Assets,
-  Button,
-  Card,
-  CardSection,
-  ChartWrapper,
-  Dropdown,
-  ErrorPage,
-  H1,
-  H2,
-  H3,
-  H4,
-  Header,
-  Icon,
-  IconSVG,
-  GlobalStyle,
-  Link,
-  Modal,
-  ModalHeading,
-  Need,
-  NeedState,
-  Search,
-  TitleXL,
-  TitleXXL,
-  fonts,
-  palette,
-  spacing,
-  zindex,
-};
+export { Assets };
 
-export * from "./components/Tabs/Tabs";
+export * from "./styles";
 
-export type { ButtonKind, ButtonProps, ModalProps };
+export * from "./components/Button";
+export * from "./components/Card";
+export * from "./components/ChartWrapper";
+export * from "./components/Dropdown";
+export * from "./components/ErrorPage";
+export * from "./components/Header";
+export * from "./components/Icon";
+export * from "./components/Modal";
+export * from "./components/Need";
+export * from "./components/Search";
+export * from "./components/Tabs";
+export * from "./components/Typography";
